### 2018-07-24 #844 @whedon
[PRE REVIEW]: Overlapping: a R package for Estimating Overlapping in Empirical Distributions
**Submitting author:** @masspastore (<a href="http://orcid.org/0000-0002-7922-6365">MASSIMILIANO PASTORE</a>)
**Repository:** <a href="https://github.com/masspastore/overlapping" target ="_blank">https://github.com/masspastore/overlapping</a>
**Version:** v1.5.0

Thanks for submitting your paper to JOSS @https://github.com/masspastore. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
