### 2018-07-24 #844 @masspastore
Dear Colleagues,

Thanks for the prompt feedback. From the list that you sent me I have
identified as possible reviewers the following based on their area of
expertise

- lebebr01; lebebr01@gmail.com

- russellpierce; russell.s.pierce@gmail.com

- jkarreth; jkarreth@ursinus.edu

Best wishes,
Massimiliano Pastore

________________________________________________________
Massimiliano Pastore
Department of Developmental and Social Psychology
Room 03-019 - Phone +39.049.827.7481
Via Venezia 8 - 35131 Padova - ITALY
http://147.162.146.188/~pastore/
