### 2018-08-18 #844 @leeper
Hi @masspastore, can you clarify: is this a method you have developed in addition to the software implementation thereof? If so, how is it related to (or distinct from) [equivalence testing](https://en.wikipedia.org/wiki/Equivalence_test)? Further, if it is a new method, you may want to expand the paper to state this explicitly as well as define some of the properties of your method in greater depth.
