+++ paper.bib
@@ -60,3 +60,83 @@
   year={2018},
   publisher={Nature Publishing Group}
 }
+
+@book{gini+livada:1943,
+	author = "Gini, C and Livada, G",
+	publisher = "Atti della {VI} Riunione della Societ{\`a} Italiana di Statistica, Roma, gennaio 1943",
+	title = "{Nuovi contributi alla teoria della transvariazione}",
+	year = "1943"
+}
+@book{weitzman:1970,
+	author = "Weitzman, Murray S",
+	publisher = "US Bureau of the Census",
+	title = "{Measures of overlap of income distributions of white and Negro families in the United States}",
+	volume = "22",
+	year = "1970"
+}
+
+@article{moravec:1988,
+	author = "Moravec, Hans P",
+	journal = "AI magazine",
+	number = "2",
+	pages = "61",
+	title = "{Sensor fusion in certainty grids for mobile robots}",
+	volume = "9",
+	year = "1988"
+}
+
+@article{viola+wells:1997,
+	author = "Viola, Paul and {Wells III}, William M",
+	journal = "International journal of computer vision",
+	number = "2",
+	pages = "137--154",
+	publisher = "Springer",
+	title = "{Alignment by maximization of mutual information}",
+	volume = "24",
+	year = "1997"
+}
+
+@article{inman+bradley:1989,
+	author = "Inman, Henry F and {Bradley Jr}, Edwin L",
+	journal = "Communications in Statistics-Theory and Methods",
+	number = "10",
+	pages = "3851--3874",
+	publisher = "Taylor \& Francis",
+	title = "{The overlapping coefficient as a measure of agreement between probability distributions and point estimation of the overlap of two normal densities}",
+	volume = "18",
+	year = "1989"
+}
+
+@book{milanovic+yitzhaki:2001,
+	author = "Milanovic, Branko and Yitzhaki, Shlomo",
+	publisher = "The World Bank",
+	title = "{Decomposing world income distribution: Does the world have a middle class?}",
+	year = "2001"
+}
+@book{cohen:1988,
+	author = "Cohen, Jacob",
+	publisher = "Psychology Press",
+	title = "{Statistical power analysis for the behavioral sciences}",
+	year = "1988"
+}
+@article{mcgraw+wong:1992,
+	author = "McGraw, Kenneth O and Wong, SP",
+	journal = "Psychological bulletin",
+	number = "2",
+	pages = "361",
+	publisher = "American Psychological Association",
+	title = "{A common language effect size statistic.}",
+	volume = "111",
+	year = "1992"
+}
+
+@article{huberty+lowman:2000,
+	author = "Huberty, Carl J and Lowman, Laureen L",
+	journal = "Educational and Psychological Measurement",
+	number = "4",
+	pages = "543--563",
+	publisher = "Sage Publications Sage CA: Thousand Oaks, CA",
+	title = "{Group overlap as a basis for effect size}",
+	volume = "60",
+	year = "2000"
+}
