### 2018-08-20 #844 @masspastore
Hi @leeper, thanks for your questions.

Actually the overlapping package presented in this paper is not a new method, but a package for easily computing an index already used in behavioral research, i.e. the overlapping index.

For making this more evident, I added a brief description and some references in the summary of the paper. In my opinion the overlapping index is not directly related to equivalence test, but like this latter it is related in some way to effect sizes measures (e.g. Cohen's d or McGraw and Wong’s CL measure).

If you think that could be interesting, I could add a brief section with a formal description of the overlapping index.
