### 2018-08-20 d1e633a
cleanup paper metadata for JOSS
