### 2018-08-20 #844 @leeper
Thanks, that's very helpful. I've sent you a quick PR to cleanup some of the paper metadata: https://github.com/masspastore/overlapping/pull/1

I'll regenerate the PDF once that's accepted.
