### 2018-08-21 #844 @leeper
@rmhogervorst Would you be interested in reviewing this for JOSS?
