### 2018-09-25 #844 @masspastore
Thanks @arfon, I would suggest reviewers with R as their primary language, and with topic areas in statistics and social sciences. Some suggestions:

- lebebr01; lebebr01@gmail.com
- russellpierce; russell.s.pierce@gmail.com
- jkarreth; jkarreth@ursinus.edu
- briatte; francois.briatte@sciencespo.fr
- soodoku; gsood07@gmail.com
