### 2018-09-25 #844 @arfon
@russellpierce - definitely! The review is a checklist-based process (not visible yet here). See this for example: https://github.com/openjournals/joss-reviews/issues/974
