### 2018-10-15 #1023 @whedon
[REVIEW]: Overlapping: a R package for Estimating Overlapping in Empirical Distributions
**Submitting author:** @masspastore (<a href="http://orcid.org/0000-0002-7922-6365">MASSIMILIANO PASTORE</a>)
**Repository:** <a href="https://github.com/masspastore/overlapping" target ="_blank">https://github.com/masspastore/overlapping</a>
**Version:** v1.5.0
