### 2018-10-15 #1023 @arfon
@soodoku, @russellpierce - please carry out your reviews in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
