### 2018-10-15 #1023 @russellpierce
Is it typical that the Statement of Need, Example, etc would be just in the paper and not in the README.md of the repo?
