### 2018-10-15 #1023 @russellpierce
@masspator Nice to meet you.  I've added several issues to your repo to cover items that I believe are in this checklist.  Rather than giving you a lot of extra feedback at this time, I'll pause there and wait to see if that feedback seems useful. :)
