### 2018-10-16 14e71b4
put paper under paper, do linting
