+++ R/overlap.R
@@ -1,103 +1,113 @@
-#rm(list=ls())
-#source("/home/el/lavori/Rdevel/overlapping_1.4.0/R/cutnumeric.R")
-#rr <- rnorm(10)
-#x <- list(X1=rr,X2=rt(10,8),X3=rchisq(10,2),X4=rr)
-#nbins <- 20; i1 <- 1; i2 <- i1+1
-#load("~/lavori/Rdevel/overlappingTest/provaoverlap.rda")
-overlap <- function(x,nbins=1000,plot=FALSE,partial.plot=FALSE) {
-  if (is.null(names(x))) names(x) <- paste("Y",1:length(x),sep="")
-  dd <- maxX <- maxY <- NULL
+#' Overlap: Estimate Overlap Between Two Distribution
+#' 
+#' @param x numeric vector
+#' @param nbins number of bins. default = 1000
+#' @param plot  boolean. default is false. 
+#' @param partial.plot boolean. default is false
+#' 
+overlap <- function(x, nbins = 1000, plot = FALSE, partial.plot = FALSE) {
+  if (is.null(names(x))) names(x) <- paste("Y", 1:length(x), sep = "")
+  dd <- maxx <- maxy <- NULL
   ## density estimation
   for (j in 1:length(x)) {
-    dj <- density(x[[j]],n=nbins)
-    maxXj <- dj$x[which(dj$y==max(dj$y))]
-    maxYj <- max(dj$y) 
-    ddd <- data.frame(x=dj$x,y=dj$y,j=names(x)[j])
-    dd <- rbind(dd,ddd)
-    maxX <- c(maxX,maxXj)
-    maxY <- c(maxY,maxYj)
+    dj <- density(x[[j]], n = nbins)
+    maxxj <- dj$x[which(dj$y == max(dj$y))]
+    maxyj <- max(dj$y)
+    ddd <- data.frame(x = dj$x, y = dj$y, j = names(x)[j])
+    dd <- rbind(dd, ddd)
+    maxx <- c(maxx, maxxj)
+    maxy <- c(maxy, maxyj)
   }
-  dd$xclass <- cut(dd$x,seq(min(dd$x),max(dd$x),length=nbins),include.lowest=TRUE)
+  dd$xclass <- cut(dd$x,
+                  seq(min(dd$x), max(dd$x), length = nbins),
+                  include.lowest = TRUE)
   dd$xnum <- cutnumeric(dd$x)
-  
+
   OV <- DD <- xpoints <- NULL
-  for (i1 in 1:(length(x)-1)) {
-    for (i2 in (i1+1):(length(x))) {
-      (Do <- order(maxX))
-      (Dx <- sort(maxX))
-      (Dy <- maxY[order(maxX)])
-      
-      (max1 <- Dx[i1])
-      (max2 <- Dx[i2])
-      #d1 <- dd[dd$j==Do[i1],]
-      #d2 <- dd[dd$j==Do[i2],]
-      (d1 <- dd[dd$j==names(x)[Do[i1]],])
-      (d2 <- dd[dd$j==names(x)[Do[i2]],])
-      
-      if (max1>max2) {
+  for (i1 in 1:(length(x) - 1)) {
+    for (i2 in (i1 + 1):(length(x))) {
+      Do <- order(maxx)
+      Dx <- sort(maxx)
+      Dy <- maxy[order(maxx)]
+
+      max1 <- Dx[i1]
+      max2 <- Dx[i2]
+      d1 <- dd[dd$j == names(x)[Do[i1]], ]
+      d2 <- dd[dd$j == names(x)[Do[i2]], ]
+
+      if (max1 > max2) {
         tram <- d2; d2 <- d1; d1 <- tram
         tram <- max2; max2 <- max1; max1 <- tram
       }
[…]
