### 2018-10-21 #844 @briatte
Dear all ~

The entire review process went too quick for me to accept the offer, but I would have accepted it gladly.

Please feel free to send social science-y packages my way in the future. The email address listed earlier on is correct.

All the best, ~ François
