+++ paper/overlap.bib
@@ -1,0 +1,315 @@
+@comment{x-kbibtex-personnameformatting=<%l><, %f>}
+
+@book{ggplot2,
+	author = "Wickham, Hadley",
+	isbn = "978-0-387-98140-6",
+	publisher = "Springer-Verlag New York",
+	title = "{ggplot2: Elegant Graphics for Data Analysis}",
+	url = "http://ggplot2.org",
+	year = "2009"
+}
+
+@book{cohen:1988,
+	author = "Cohen, Jacob",
+	publisher = "Psychology Press",
+	title = "{Statistical power analysis for the behavioral sciences}",
+	year = "1988"
+}
+
+@article{nakagawa+kuthil:2007,
+	author = "Nakagawa, Shinichi and Cuthill, Innes C",
+	journal = "Biological reviews",
+	number = "4",
+	pages = "591--605",
+	publisher = "Wiley Online Library",
+	title = "{Effect size, confidence interval and statistical significance: a practical guide for biologists}",
+	volume = "82",
+	year = "2007"
+}
+
+@article{kelley+preacher:2011,
+	author = "Preacher, Kristopher J and Kelley, Ken",
+	journal = "Psychological methods",
+	number = "2",
+	pages = "93",
+	publisher = "American Psychological Association",
+	title = "{Effect size measures for mediation models: quantitative strategies for communicating indirect effects.}",
+	volume = "16",
+	year = "2011"
+}
+
+@inproceedings{goldberg+etal:2010,
+	author = "Goldberg, Mark K and Hayvanovych, Mykola and Magdon-Ismail, Malik",
+	booktitle = "{Social Computing (SocialCom), 2010 {IEEE} Second International Conference on}",
+	organization = "IEEE",
+	pages = "303--308",
+	title = "{Measuring similarity between sets of overlapping clusters}",
+	year = "2010"
+}
+
+@book{gini+livada:1943,
+	author = "Gini, C and Livada, G",
+	publisher = "Atti della {VI} Riunione della Societ{\`a} Italiana di Statistica, Roma, gennaio 1943",
+	title = "{Nuovi contributi alla teoria della transvariazione}",
+	year = "1943"
+}
+
+@book{weitzman:1970,
+	author = "Weitzman, Murray S",
+	publisher = "US Bureau of the Census",
+	title = "{Measures of overlap of income distributions of white and Negro families in the United States}",
+	volume = "22",
+	year = "1970"
+}
+
+@article{moravec:1988,
+	author = "Moravec, Hans P",
+	journal = "AI magazine",
+	number = "2",
+	pages = "61",
+	title = "{Sensor fusion in certainty grids for mobile robots}",
+	volume = "9",
+	year = "1988"
+}
+
+@article{viola+wells:1997,
+	author = "Viola, Paul and {Wells III}, William M",
+	journal = "International journal of computer vision",
+	number = "2",
+	pages = "137--154",
+	publisher = "Springer",
+	title = "{Alignment by maximization of mutual information}",
+	volume = "24",
+	year = "1997"
+}
+
+@article{inman+bradley:1989,
+	author = "Inman, Henry F and {Bradley Jr}, Edwin L",
+	journal = "Communications in Statistics-Theory and Methods",
+	number = "10",
+	pages = "3851--3874",
+	publisher = "Taylor \& Francis",
+	title = "{The overlapping coefficient as a measure of agreement between probability distributions and point estimation of the overlap of two normal densities}",
+	volume = "18",
+	year = "1989"
+}
+
+@book{milanovic+yitzhaki:2001,
+	author = "Milanovic, Branko and Yitzhaki, Shlomo",
+	publisher = "The World Bank",
[…]
