+++ paper/paper.bib
@@ -140,3 +140,14 @@
 	volume = "60",
 	year = "2000"
 }
+
+@Article{ridout+linkie:2009,
+  title = {Estimating overlap of daily activity patterns from camera trap data},
+  author = {Martin Ridout and Matthew Linkie},
+  journal = {Journal of Agricultural, Biological, and Environmental Statistics},
+  year = {2009},
+  volume = {14},
+  number = {3},
+  pages = {322--337},
+}
+
