### 2018-10-25 183909c
Update paper.md
