### 2018-10-25 483f6b3
Update paper.md
