+++ paper/paper.md
@@ -29,7 +29,7 @@ Overlapping can be defined as the area intersected by two or more probability de
 
 The package is available from GitHub (https://github.com/masspastore/overlapping) and CRAN (https://cran.r-project.org/package=overlapping). A full reference manual can be found at https://cran.r-project.org/web/packages/overlapping/overlapping.pdf.
 
-A recent R package ``overlap`` [@ridout+linkie:2009] offers an implementation of the overlapping index which can be used to analyse temporal activity patterns of animals and species in echology. However, ``overlapping`` package offers a more general approach where overlapping can be computed for any type of variable, allowing for computations with more than two variables.
+A recent R package, ``overlap`` [@ridout+linkie:2009], offers an implementation of the overlapping index which can be used to analyse temporal activity patterns of animals and species in echology. However, ``overlapping`` package offers a more general approach where overlapping can be computed for any type of variable, allowing for computations with more than two variables.
 
 
 # Examples
