+++ paper/paper.md
@@ -3,11 +3,11 @@ title: 'Overlapping: a R package for Estimating Overlapping in Empirical Distrib
 authors:
 - name: Massimiliano Pastore
 - affiliation: 1
-date: "23 july 2018"
+date: "25 October 2018"
 output:
+  pdf_document: default
   keep_tex: yes
   self_contained: no
-  pdf_document: default
   html_document:
     df_print: paged
   word_document: default
@@ -29,7 +29,7 @@ Overlapping can be defined as the area intersected by two or more probability de
 
 The package is available from GitHub (https://github.com/masspastore/overlapping) and CRAN (https://cran.r-project.org/package=overlapping). A full reference manual can be found at https://cran.r-project.org/web/packages/overlapping/overlapping.pdf.
 
-A recent R package, ``overlap`` [@ridout+linkie:2009], offers an implementation of the overlapping index which can be used to analyse temporal activity patterns of animals and species in echology. However, ``overlapping`` package offers a more general approach where overlapping can be computed for any type of variable, allowing for computations with more than two variables.
+A recent R, package ``overlap`` [@ridout+linkie:2009], offers an implementation of the overlapping index which can be used to analyse temporal activity patterns of animals and species in echology. However, ``overlapping`` package offers a more general approach where overlapping can be computed for any type of variable, allowing for computations with more than two variables.
 
 
 # Examples
