### 2018-10-25 #1023 @arfon
@masspastore - please update your paper YAML header to _exactly_ match the format of this example: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

Please note that YAML is whitespace sensitive.
