+++ paper/paper.bib
@@ -141,7 +141,7 @@
 	year = "2000"
 }
 
-@Article{ridout+linkie:2009,
+@article{ridout+linkie:2009,
   title = {Estimating overlap of daily activity patterns from camera trap data},
   author = {Martin Ridout and Matthew Linkie},
   journal = {Journal of Agricultural, Biological, and Environmental Statistics},
