### 2018-10-26 #1023 @masspastore
Thanks @arfon

I unfortunately added a new reference in the bib file without checking it.
I updated the bib and the paper files, now I think it is correct.
