### 2018-10-27 bed721b
Update paper.md
