### 2018-10-27 2b06209
Update paper.md
