### 2018-10-27 f7af282
Update paper.md
