### 2018-10-27 #1023 @arfon
@masspastore - please merge this PR too: https://github.com/masspastore/overlapping/pull/9
