### 2018-11-05 #1023 @russellpierce
Oh darn, @soodoku I might have checked items off your list when I meant to be checking off mine.  My apologies!
