### 2018-11-05 #1023 @soodoku
No worries.

On Mon, Nov 5, 2018 at 7:53 AM Russell S. Pierce <notifications@github.com>
wrote:

> Oh darn, @soodoku <https://github.com/soodoku> I might have checked items
> off your list when I meant to be checking off mine. My apologies!
>
