+++ man/overlap.Rd
@@ -8,7 +8,7 @@ Overlapping estimation
 It gives the overlapped estimated area of two or more empirical distributions. %Stima l'area di sovrapposizione tra distribuzioni empiriche. 
 }
 \usage{
-overlap(x, nbins = 1000, plot = FALSE, partial.plot = FALSE)
+overlap(x, nbins = 1000, plot = FALSE, partial.plot = FALSE, ... )
 }
 %- maybe also 'usage' for other objects documented here.
 \arguments{
@@ -16,6 +16,7 @@ overlap(x, nbins = 1000, plot = FALSE, partial.plot = FALSE)
   \item{nbins}{number of equally spaced points at which the overlapping density is evaluated}
   \item{plot}{logical, if \code{TRUE}, final plot of estimated densities and overlapped areas is produced}
   \item{partial.plot}{logical, if \code{TRUE}, partial paired distributions are plotted}
+  \item{...}{type of kernel, see function \code{\link{density}}}
 }
 \details{
 If the list \code{x} contains more than two elements (i.e. more than two distributions) it computes all overlapping between all paired distributions. Partial plots refer to these coupled distributions. 
