+++ paper/paper.bib
@@ -1,24 +1,276 @@
+% Encoding: UTF-8
 @comment{x-kbibtex-personnameformatting=<%l><, %f>}
 
-@book{ggplot2,
-	author = "Wickham, Hadley",
-	isbn = "978-0-387-98140-6",
-	publisher = "Springer-Verlag New York",
-	title = "{ggplot2: Elegant Graphics for Data Analysis}",
-	url = "http://ggplot2.org",
-	year = "2009"
+@Book{ggplot2,
+  title     = {{ggplot2: Elegant Graphics for Data Analysis}},
+  publisher = {Springer-Verlag New York},
+  year      = {2009},
+  author    = {Wickham, Hadley},
+  isbn      = {978-0-387-98140-6},
+  doi       = {10.1007/978-0-387-98141-3_6},
+  url       = {http://ggplot2.org},
 }
 
+@Book{cohen:1988,
+  title     = {{Statistical power analysis for the behavioral sciences}},
+  publisher = {Psychology Press},
+  year      = {1988},
+  author    = {Cohen, Jacob},
+  isbn      = {0-8058-0283-5},
+}
 
-@article{marci+al:2018,
-	author = "Marci, Tatiana and Lionetti, Francesca and Moscardino, Ughetta and Pastore, Massimiliano and Calvo, Vincenzo and Alto{\'e}, Gianmarco",
-	journal = "European Journal of Developmental Psychology",
+@article{nakagawa+kuthil:2007,
+	author = "Nakagawa, Shinichi and Cuthill, Innes C",
+	journal = "Biological reviews",
 	number = "4",
-	pages = "481--492",
+	pages = "591--605",
+	publisher = "Wiley Online Library",
+	title = "{Effect size, confidence interval and statistical significance: a practical guide for biologists}",
+	volume = "82",
+	year = "2007"
+}
+
+@article{kelley+preacher:2011,
+	author = "Preacher, Kristopher J and Kelley, Ken",
+	journal = "Psychological methods",
+	number = "2",
+	pages = "93",
+	publisher = "American Psychological Association",
+	title = "{Effect size measures for mediation models: quantitative strategies for communicating indirect effects.}",
+	volume = "16",
+	year = "2011"
+}
+
+@inproceedings{goldberg+etal:2010,
+	author = "Goldberg, Mark K and Hayvanovych, Mykola and Magdon-Ismail, Malik",
+	booktitle = "{Social Computing (SocialCom), 2010 {IEEE} Second International Conference on}",
+	organization = "IEEE",
+	pages = "303--308",
+	title = "{Measuring similarity between sets of overlapping clusters}",
+	year = "2010"
+}
+
+@book{gini+livada:1943,
+	author = "Gini, C and Livada, G",
+	publisher = "Atti della {VI} Riunione della Societ{\`a} Italiana di Statistica, Roma, gennaio 1943",
+	title = "{Nuovi contributi alla teoria della transvariazione}",
+	year = "1943"
+}
+
+@book{weitzman:1970,
+	author = "Weitzman, Murray S",
+	publisher = "US Bureau of the Census",
+	title = "{Measures of overlap of income distributions of white and Negro families in the United States}",
+	volume = "22",
+	year = "1970"
+}
+
+@Article{moravec:1988,
+  author  = {Moravec, Hans P},
+  title   = {{Sensor fusion in certainty grids for mobile robots}},
+  journal = {AI magazine},
+  year    = {1988},
+  volume  = {9},
+  number  = {2},
+  pages   = {61},
+  doi     = {10.1609/aimag.v9i2.676},
+}
+
+@Article{viola+wells:1997,
+  author    = {Viola, Paul and {Wells III}, William M},
[…]
