### 2018-11-06 06dfb86
Update paper.md
