+++ paper/paper.bib
@@ -273,11 +273,13 @@
   publisher = {Taylor \& Francis},
 }
 
-@unpublished{altoe+al:inpress,
-	author = "Alto{\`e}, Gianmarco and D'Amore, Giuseppe and Scalfari, Francesco",
-	note = "in press",
-	year = "in press",
-	title = "{Skulls and transvariation}"
+@InCollection{altoe+al:2018,
+  author    = {Alto{\`e}, Gianmarco and D'Amore, Giuseppe and Scalfari, Francesco},
+  title     = {{Skulls and transvariation}},
+  booktitle = {Biostat at 25 Invited Essays in Theoretical, Biomedical and Social Statistics},
+  publisher = {Edizioni ETS, PISA},
+  year      = {2018},
+  editor    = {Mario Di Bacco and Francesco Scalfari},
 }
 
 @Article{pluess+al:2018,
