+++ R/overlap.R
@@ -4,113 +4,78 @@
 #' @param nbins number of bins. default = 1000
 #' @param plot  boolean. default is false. 
 #' @param partial.plot boolean. default is false
-#' @param ... type of kernel see function density
-overlap <- function(x, nbins = 1000, plot = FALSE, partial.plot = FALSE, ... ) {
+#' @param boundaries an optional list 
+#' @param ... options see function density
+#' 
+overlap <- function(x, nbins = 1000, plot = FALSE, 
+                    partial.plot = FALSE, boundaries = NULL, ... ) {
+
   if (is.null(names(x))) names(x) <- paste("Y", 1:length(x), sep = "")
-  dd <- maxX <- maxY <- NULL
+  dd <- OV <- FUNC <- DD <- xpoints <- COMPTITLE <- NULL
   
   ## density estimation
   for (j in 1:length(x)) {
-    dj <- density(x[[j]], n = nbins, bw = "nrd0", 
-                  kernel = c("gaussian", "epanechnikov", "rectangular",
-                            "triangular", "biweight",
-                            "cosine", "optcosine") )
-    maxXj <- dj$x[which(dj$y == max(dj$y))]
-    maxYj <- max(dj$y) 
-    ddd <- data.frame(x = dj$x, y = dj$y, j = names(x)[j])
-    dd <- rbind(dd, ddd)
-    maxX <- c(maxX, maxXj)
-    maxY <- c(maxY, maxYj)
-  }
-  dd$xclass <- cut(dd$x, seq(min(dd$x), max(dd$x), length = nbins),
-                   include.lowest = TRUE)
-  dd$xnum <- cutnumeric(dd$x, n = nbins)
-  
-  OV <- DD <- xpoints <- NULL
-  for (i1 in 1:(length(x)-1)) {
-    for (i2 in (i1+1):(length(x))) {
-      (Do <- order(maxX))
-      (Dx <- sort(maxX))
-      (Dy <- maxY[order(maxX)])
+    
+    ## boundaries check
+    if (!is.null(boundaries)) {
       
-      (max1 <- Dx[i1])
-      (max2 <- Dx[i2])
-      (d1 <- dd[dd$j == names(x)[Do[i1]], ])
-      (d2 <- dd[dd$j == names(x)[Do[i2]], ])
-      
-      if (max1 > max2) {
-        tram <- d2; d2 <- d1; d1 <- tram
-        tram <- max2; max2 <- max1; max1 <- tram
-      }
-      YLIM <- range(c(d1$y, d2$y))
-      XLIM <- range(c(d1$xnum, d2$xnum))
-      
-      XNUM <- unique(dd$xnum)
-      dominance <- rep(NA, length(XNUM))
-      change <- NULL
-      
-      for (h in 1:length(XNUM)) {
-        if (length(d1$y[d1$xnum == XNUM[h]]) > 0) {
-          Y1 <- max(d1$y[d1$xnum == XNUM[h]]) # possible warnings here          
-        } else {
-          Y1 <- -Inf
-        }
-        if (length(d2$y[d2$xnum == XNUM[h]]) > 0) {
-          Y2 <- max(d2$y[d2$xnum == XNUM[h]]) # possible warnings here  
-        } else {
-          Y2 <- -Inf
-        }
-        
-        dominance[h] <- ifelse(Y1>Y2, 1, 2)
-        if (h > 1) {
-          if (dominance[h] != dominance[h-1]) change <- c(change, h-1)
+      Lbound <- lapply(boundaries,FUN=length)
[…]
