+++ man/cutnumeric.Rd
@@ -19,7 +19,7 @@ cutnumeric(x, n = 1000)
   \item{n}{number of classes}
 }
 \details{
-It calls the \code{cut} function, and then converts factor classess in numeric classes, returning for each class its central value. 
+It calls the \code{cut} function, and then converts factor classes in numeric classes, returning for each class its central value. 
 %e quindi converte le classi ottenute (di tipo factor) in classi numeriche assegnando il valore centrale dell'intervallo. 
 }
 \value{
@@ -31,9 +31,9 @@ It returns a numerical vector. The values are the central points of classes obta
 \author{
 Massimiliano Pastore
 }
-\note{
-This function is called from the function \code{\link{overlap}}.
-}
+%\note{
+%This function is called from the function \code{\link{overlap}}.
+%}
 
 %% ~Make other sections like Warning with \section{Warning }{....} ~
 
