+++ man/overlap.Rd
@@ -8,7 +8,8 @@ Overlapping estimation
 It gives the overlapped estimated area of two or more empirical distributions. %Stima l'area di sovrapposizione tra distribuzioni empiriche. 
 }
 \usage{
-overlap(x, nbins = 1000, plot = FALSE, partial.plot = FALSE, ... )
+overlap(x, nbins = 1000, plot = FALSE, 
+    partial.plot = FALSE, boundaries = NULL, ... )
 }
 %- maybe also 'usage' for other objects documented here.
 \arguments{
@@ -16,29 +17,28 @@ overlap(x, nbins = 1000, plot = FALSE, partial.plot = FALSE, ... )
   \item{nbins}{number of equally spaced points at which the overlapping density is evaluated}
   \item{plot}{logical, if \code{TRUE}, final plot of estimated densities and overlapped areas is produced}
   \item{partial.plot}{logical, if \code{TRUE}, partial paired distributions are plotted}
-  \item{...}{type of kernel, see function \code{\link{density}}}
+  \item{boundaries}{an optional list for bounded distributions, see Details}
+  \item{...}{optional arguments to be passed to function \code{\link{density}}}
 }
 \details{
-If the list \code{x} contains more than two elements (i.e. more than two distributions) it computes all overlapping between all paired distributions. Partial plots refer to these coupled distributions. 
-%Se la lista \code{x} contiene piu di due elementi, calcola le sovrapposizioni tra tutte le coppie di distribuzioni. I grafici parziali si riferiscono a queste coppie.
+If the list \code{x} contains more than two elements (i.e. more than two distributions) it computes all overlapping between  paired distributions. Partial plots refer to these paired distributions. 
 
-If \code{plot=TRUE}, all overlapped areas are plotted.  It requires \code{lattice}.
-%Se si pone a \code{plot=TRUE} richiede \code{lattice} e produce un grafico complessivo delle aree di sovrapposizione rispetto ciascuna coppia.
+If \code{plot=TRUE}, all overlapped areas are plotted.  It requires \code{ggplot2}.
+
+The optional list \code{boundaries} must contain two elements: \code{from} and \code{to}, indicating the empirical limits of input variables. Each element must be of length equal to the input data list \code{x} or, at least, length one when all boundaries are equal for all distributions. See Examples below. 
+
 }
 \value{
 It returns a list containing the following components:
   \item{DD }{Data frame with information used for computing overlapping, containing the following variables: 
   \code{x}, coordinates of the points where the density is estimated;
-  \code{y}, density;
-  \code{j}, index of the distribution in the list \code{x};
-  \code{xclass}, class of \code{x};
-  \code{xnum}, numerical class of \code{x} (obtained by \code{\link{cutnumeric}});
+  \code{y1} and \code{y2}, densities;
+  \code{ovy}, density for estimating overlapping area (i.e. \code{min(y1,y2)});
+  \code{ally}, density for estimating whole area (i.e. \code{max(y1,y2)});
   \code{dominance}, indicates which distribution has the highest density;
-  \code{w}, flag 0-1 for normalizing area;
-  \code{k}, label indicating which distributions are compared %etichetta per il confronto, per \code{default} viene ricavato dai nomi degli oggetti nella lista di input.
-  }
-  \item{OV }{Estimates of overlapped areas relative to each couple of distributions.}
-  \item{xpoints }{Abscissas of intersection points among the density curves.}
[…]
