+++ tests/testthat/test-overlap.R
@@ -4,91 +4,30 @@ test_that("overlap works", {
   set.seed(20150605)
   x <- list(X1=rnorm(100),X2=rt(50,8))
   nbins <- 1000
-  dd <- maxX <- maxY <- NULL
+  dd <- FUNC <- OV <- COMPTITLE <- NULL
   
-  ## density estimation
   for (j in 1:length(x)) {
     dj <- density(x[[j]], n = nbins)
-    maxXj <- dj$x[which(dj$y == max(dj$y))]
-    maxYj <- max(dj$y) 
-    ddd <- data.frame(x = dj$x, y = dj$y, j = names(x)[j])
+    ddd <- data.frame(x = dj$x, y = dj$y, j = names(x)[j]) 
+    FUNC <- c(FUNC, list(with(ddd,approxfun(x,y))))
     dd <- rbind(dd, ddd)
-    maxX <- c(maxX, maxXj)
-    maxY <- c(maxY, maxYj)
   }
-  dd$xclass <- cut(dd$x, seq(min(dd$x), max(dd$x), length = nbins),
-                   include.lowest = TRUE)
-  ### cutnumeric
-  xclass <- cut(dd$x, seq( min(dd$x), max(dd$x), length = 1000 ), include.lowest = TRUE )
-  dd$xnum <- unlist( lapply( xclass, function(b) {
-    h <- strsplit( as.character(b), "," )[[1]]
-    h <- as.numeric(gsub( "\\[", "", 
-                          gsub("\\]", "", gsub("\\(", "", h ) ) ))
-    mean( h )
-  } ) )
   
-  ## overlap estimation
-  OV <- DD <- xpoints <- NULL
   for (i1 in 1:(length(x)-1)) {
     for (i2 in (i1+1):(length(x))) {
-      (Do <- order(maxX))
-      (Dx <- sort(maxX))
-      (Dy <- maxY[order(maxX)])
+      comptitle <- paste0(names(x)[i1],"-",names(x)[i2])
       
-      (max1 <- Dx[i1])
-      (max2 <- Dx[i2])
-      (d1 <- dd[dd$j == names(x)[Do[i1]], ])
-      (d2 <- dd[dd$j == names(x)[Do[i2]], ])
+      dd2 <- data.frame(x=dd$x,y1=FUNC[[i1]](dd$x),y2=FUNC[[i2]](dd$x))    
+      dd2[is.na(dd2)] <- 0
+      dd2$ovy <- apply(dd2[,-1],1,min)
+      dd2$ally <- apply(dd2[,c("y1","y2")],1,max,na.rm=TRUE)
+      dd2$dominance <- with(dd2, ifelse(y1>y2,1,2))
+      dd2$k <- comptitle
       
-      if (max1 > max2) {
-        tram <- d2; d2 <- d1; d1 <- tram
-        tram <- max2; max2 <- max1; max1 <- tram
-      }
-      YLIM <- range(c(d1$y, d2$y))
-      XLIM <- range(c(d1$xnum, d2$xnum))
-      
-      XNUM <- unique(dd$xnum)
-      dominance <- rep(NA, length(XNUM))
-      change <- NULL
-      
-      for (h in 1:length(XNUM)) {
-        if (length(d1$y[d1$xnum == XNUM[h]]) > 0) {
-          Y1 <- max(d1$y[d1$xnum == XNUM[h]]) # possible warnings here          
-        } else {
-          Y1 <- -Inf
-        }
-        if (length(d2$y[d2$xnum == XNUM[h]]) > 0) {
-          Y2 <- max(d2$y[d2$xnum == XNUM[h]]) # possible warnings here  
-        } else {
[…]
