### 2018-11-19 #1023 @arfon
@soodoku, @russellpierce - it looks like you're both making good progress with your reviews here. Do you think you might be able to complete them in the next week?
