+++ paper/paper.bib
@@ -328,4 +328,16 @@
     pages = {322--337},
   }
 
+
+@Article{wilkinson:1999,
+	author = "Wilkinson, L and {Task Force on Statistical Inference}",
+	journal = {American Psychologist},
+	number = "8",
+	pages = {594--604},
+	title = {Statistical methods in psychology journals - Guidelines and explanations},
+	volume = {54},
+	year = {1999}
+}
+
+
 @Comment{jabref-meta: databaseType:bibtex;}
