+++ paper/paper.md
@@ -1,35 +1,35 @@
 ---
 title: 'Overlapping: a R package for Estimating Overlapping in Empirical Distributions'
-tags:
-  - R
-  - statistics
 authors:
-  - name: Massimiliano Pastore
-    affiliation: 1
+- affiliation: 1
+  name: Massimiliano Pastore
+date: "25 October 2018"
+bibliography: paper.bib
+tags:
+- R
+- statistics
 affiliations:
-  - name: Department of Developmental and Social Psychology, University of Padova
-    index: 1
-date: 25 October 2018
-bibliography: paper.bib 
+- index: 1
+  name: Department of Developmental and Social Psychology, University of Padova
 ---
 
 # Summary
 
-Overlapping can be defined as the area intersected by two or more probability density functions. The idea of overlapping was introduced in a formal way by @gini+livada:1943 and, more recently, it has been applied in several research problems involving, for instance, data fusion [@moravec:1988], information processing [@viola+wells:1997], applied statistics [@inman+bradley:1989], economics [@milanovic+yitzhaki:2001], psychology, as a basis for Cohen's $U$ index [@cohen:1988], McGraw and Wong's $CL$ measure [@mcgraw+wong:1992], and  Huberty's $I$ degree of non-overlap index [@huberty+lowman:2000].
+Overlapping can be defined as the area intersected by two or more probability density functions. The idea of overlapping was introduced in a formal way by @gini+livada:1943 and, more recently, it has been applied in several research problems involving, for instance, data fusion [@moravec:1988], information processing [@viola+wells:1997], applied statistics [@inman+bradley:1989], economics [@milanovic+yitzhaki:2001] and psychology, as a basis for Cohen's $U$ index [@cohen:1988], McGraw and Wong's $CL$ measure [@mcgraw+wong:1992], and  Huberty's $I$ degree of non-overlap index [@huberty+lowman:2000].
 
-``overlapping`` is a R package for estimating the overlapping area of two or more empirical distributions. The main idea of the package is to offer an easy way to quantify the similarity (or the difference) between two or more empirical distributions. In addition, the package allows to plot density distributions, highlighting the overlapped area by using the ``ggplot2`` R package [@ggplot2].
+``overlapping`` is an R package for estimating the overlapping area of two or more kernel density estimations from empirical data. The main idea of the package is to offer an easy way to quantify the similarity (or the difference) between two or more empirical distributions. In addition, the package allows to plot density distributions, highlighting the overlapped area by using the ``ggplot2`` R package [@ggplot2].
 
 
 The package is available from GitHub (https://github.com/masspastore/overlapping) and CRAN (https://cran.r-project.org/package=overlapping). A full reference manual can be found at https://cran.r-project.org/web/packages/overlapping/overlapping.pdf.
 
-A recent R package, ``overlap`` [@ridout+linkie:2009], offers an implementation of the overlapping index which can be used to analyse temporal activity patterns of animals and species in echology. However, ``overlapping`` package offers a more general approach where overlapping can be computed for any type of variable, allowing for computations with more than two variables.
+A recent R package, ``overlap`` [@ridout+linkie:2009], offers an implementation of the overlapping index which can be used to analyse temporal activity patterns of animals and species in echology. Compared to this latter, ``overlapping`` package offers a more general approach where overlapping can be computed for any type of numerical variable, and it allows for computations with more than two variables.
 
 
 # Examples
 
-Suppose we have collected data in two groups of 100 subjects each with respect to a generic variable *Y*, expressed by scores ranging between 0 and 30, and to be interested in assessing whether the two groups can be considered samples from populations with the same average.
[…]
