### 2018-11-28 #1023 @russellpierce
@masspastore I added one more grammar correction under a git blame. Otherwise I think things are looking good.  cc: @arfon
