### 2018-11-29 #1023 @arfon
@soodoku - it looks like there are a few checkboxes un-checked for your review. Is there still work to be done by the author or are you happy to accept here?
