### 2018-11-30 #1023 @russellpierce
@soodoku thanks for giving it a thorough look. I'll admit that I've been through it enough at this point where I don't have as clear as view on the end result as I might have at the start. So, it is entirely possible I'm missing something. :) I.e. don't take my signoff as a need for less rigour.
