+++ paper/paper.bib
@@ -26,6 +26,7 @@
 	pages = "591--605",
 	publisher = "Wiley Online Library",
 	title = "{Effect size, confidence interval and statistical significance: a practical guide for biologists}",
+	doi = {10.1111/j.1469-185X.2007.00027.x},
 	volume = "82",
 	year = "2007"
 }
@@ -37,6 +38,7 @@
 	pages = "93",
 	publisher = "American Psychological Association",
 	title = "{Effect size measures for mediation models: quantitative strategies for communicating indirect effects.}",
+	doi = {10.1037/a0022658},
 	volume = "16",
 	year = "2011"
 }
@@ -139,6 +141,7 @@
 	pages = "1583--1596",
 	publisher = "Elsevier",
 	title = "{Nonparametric estimation of the coefficient of overlapping---theory and empirical application}",
+	doi = {10.1016/j.csda.2005.01.014},
 	volume = "50",
 	year = "2006"
 }
@@ -170,6 +173,7 @@
 	pages = "51--61",
 	publisher = "Elsevier",
 	title = "{A nonparametric measure of the overlapping coefficient}",
+	doi = {10.1016/S0167-9473(99)00074-2},
 	volume = "34",
 	year = "2000"
 }
@@ -182,6 +186,7 @@
 	publisher = "Sage Publications Sage CA: Los Angeles, CA",
 	title = "{A critique and improvement of the CL common language effect size statistics of McGraw and Wong}",
 	volume = "25",
+	doi = {10.2307/1165329},
 	year = "2000"
 }
 
@@ -192,6 +197,7 @@
 	pages = "22--50",
 	publisher = "Taylor \& Francis",
 	title = "{Beyond {C}ohen's $d$: Alternative effect size measures for between-subject designs}",
+	doi = {10.1080/00220973.2012.745471},
 	volume = "82",
 	year = "2014"
 }
@@ -204,6 +210,7 @@
 	publisher = "Taylor \& Francis Group",
 	title = "{Standard distance in univariate and multivariate analysis}",
 	volume = "40",
+	doi = {10.2307/2684560}, 
 	year = "1986"
 }
 
@@ -325,6 +332,7 @@
     year = {2009},
     volume = {14},
     number = {3},
+    doi={10.1198/jabes.2009.08038},
     pages = {322--337},
   }
 
@@ -338,6 +346,5 @@
 	volume = {54},
 	year = {1999}
 }
-
 
 @Comment{jabref-meta: databaseType:bibtex;}
