### 2018-12-05 #1023 @masspastore
Hi @arfon, I updated the bib file with all avaliable DOIs

Thanks
